Move elements out to their own file
Amos King @adkron <amos@binarynoggin.com>
