+++ lib/openxml/elements/notes_size.rb
@@ -1,0 +1,14 @@
+require "openxml/extract/element"
+
+module OpenXml
+  module Pptx
+    module Elements
+      class NotesSize < OpenXml::Element
+        tag :notesSz
+
+        attribute :cx
+        attribute :cy
+      end
+    end
+  end
+end
