+++ lib/openxml/elements/slide_size.rb
@@ -1,0 +1,14 @@
+require "openxml/extract/element"
+
+module OpenXml
+  module Pptx
+    module Elements
+      class SlideSize < OpenXml::Element
+        tag :sldSz
+
+        attribute :cx
+        attribute :cy
+      end
+    end
+  end
+end
